app/requirements.txt
flask==3.0.0
requests==2.31.0
urllib3==2.0.7
pyyaml==6.0.1
jinja2==3.1.2
cryptography==41.0.5
